+++ packages/osd-plugin-helpers/src/integration_tests/build.test.ts
@@ -32,8 +32,8 @@ import Path from 'path';
 import Fs from 'fs';
 
 import execa from 'execa';
-import { REPO_ROOT } from '@osd/utils';
-import { createStripAnsiSerializer, createReplaceSerializer } from '@osd/dev-utils';
+import { standardize } from '@osd/cross-platform';
+import { REPO_ROOT, createStripAnsiSerializer, createReplaceSerializer } from '@osd/dev-utils';
 import extract from 'extract-zip';
 import del from 'del';
 import globby from 'globby';
@@ -78,7 +78,7 @@ it('builds a generated plugin into a viable archive', async () => {
   expect(generateProc.all).toMatchInlineSnapshot(`
     " succ 🎉
 
-          Your plugin has been created in plugins/foo_test_plugin
+          Your plugin has been created in ${standardize('plugins/foo_test_plugin', false, true)}
     "
   `);
 
@@ -165,7 +165,7 @@ it('builds a non-semver generated plugin into a viable archive', async () => {
   expect(generateProc.all).toMatchInlineSnapshot(`
     " succ 🎉
 
-          Your plugin has been created in plugins/foo_test_plugin
+          Your plugin has been created in ${standardize('plugins/foo_test_plugin', false, true)}
     "
   `);